all competencies added
