+++ src/app/client/src/app/modules/explore-page/components/explore-page/explore-page.component.scss
@@ -119,6 +119,7 @@
 .cardContainer{
   display: flex;
   justify-content: space-between;
+  cursor: pointer;
 }
 .browseByCard{
   width: 30%;
