+++ src/app/client/src/app/modules/search/components/home-search/home-search.component.scss
@@ -1,0 +1,8 @@
+@import "compass-sb-styles/assets/mixins/mixins";
+
+.all-comp-container{
+    flex-direction: column;
+}
+.allCompetencyTitle{
+    @include typography(var(--font-com-en), 24px, 600, 0.25px, 36px);
+}
